+++ tests/test_parser.py
@@ -2,7 +2,7 @@ import argparse
 import textwrap
 from enum import Enum
 from pathlib import Path
-from typing import List, Optional, Type, TypeVar
+from typing import List, Optional, Type, TypeVar, Union
 
 import pytest
 from typing_extensions import Literal
@@ -10,7 +10,12 @@ from typing_extensions import Literal
 from typed_argparse import Parser, TypedArgs, arg
 from typed_argparse.parser import Bindings
 
-from ._testing_utils import argparse_error, compare_verbose, pre_python_10
+from ._testing_utils import (
+    argparse_error,
+    compare_verbose,
+    pre_python_3_10,
+    starting_with_python_3_10,
+)
 
 T = TypeVar("T", bound=TypedArgs)
 
@@ -94,6 +99,44 @@ def test_path() -> None:
 
     args = parse(Args, ["--path", "/my/path"])
     assert args.path == Path("/my/path")
+
+
+@starting_with_python_3_10
+def test_variations_of_optionality() -> None:
+    class Args(TypedArgs):
+        traditional: Optional[int]
+        as_union_a: Union[int, None]
+        as_union_b: Union[None, int]
+        as_union_modern_a: int | None  # type: ignore[syntax, unused-ignore]
+        as_union_modern_b: None | int  # type: ignore[syntax, unused-ignore]
+
+    args = parse(Args, [])
+    assert args.traditional is None
+    assert args.as_union_a is None
+    assert args.as_union_b is None
+    assert args.as_union_modern_a is None
+    assert args.as_union_modern_b is None
+
+    args = parse(
+        Args,
+        [
+            "--traditional",
+            "1",
+            "--as-union-a",
+            "2",
+            "--as-union-b",
+            "3",
+            "--as-union-modern-a",
+            "4",
+            "--as-union-modern-b",
+            "5",
+        ],
+    )
+    assert args.traditional == 1
+    assert args.as_union_a == 2
+    assert args.as_union_b == 3
+    assert args.as_union_modern_a == 4
+    assert args.as_union_modern_b == 5
 
 
 # Positional
@@ -431,7 +474,7 @@ def test_enum__basics(use_literal_enum: bool) -> None:
     assert "argument --enum-int: invalid choice: 'c' (choose from a, b)" == str(e.error)
 
 
-@pre_python_10
+@pre_python_3_10
 def test_enum__help_text(capsys: pytest.CaptureFixture[str]) -> None:
     class StrEnum(Enum):
         a = "a-value"
@@ -779,7 +822,7 @@ def test_parser_run__typical_lazy_syntax() -> None:
 # Defaults in help text
 
 
-@pre_python_10
+@pre_python_3_10
 def test_defaults_in_help_text__on_by_default(capsys: pytest.CaptureFixture[str]) -> None:
     class Args(TypedArgs):
         epsilon: float = arg(help="Some epsilon", default=0.1)
@@ -800,7 +843,7 @@ def test_defaults_in_help_text__on_by_default(capsys: pytest.CaptureFixture[str]
     )
 
 
-@pre_python_10
+@pre_python_3_10
 def test_defaults_in_help_text__off_if_desired(capsys: pytest.CaptureFixture[str]) -> None:
     class Args(TypedArgs):
         epsilon: float = arg(help="Some epsilon", default=0.1, auto_default_help=False)
@@ -824,7 +867,7 @@ def test_defaults_in_help_text__off_if_desired(capsys: pytest.CaptureFixture[str
 # Support of formatter class in help texts
 
 
-@pre_python_10
+@pre_python_3_10
 def test_formatter_class_support(capsys: pytest.CaptureFixture[str]) -> None:
     class Args(TypedArgs):
         foo: int = arg(help="arg line1\narg line2")
@@ -865,7 +908,7 @@ def test_formatter_class_support(capsys: pytest.CaptureFixture[str]) -> None:
 # Custom names for meta vars
 
 
-@pre_python_10
+@pre_python_3_10
 def test_metavars_in_help_text(capsys: pytest.CaptureFixture[str]) -> None:
     class Args(TypedArgs):
         epsilon: float = arg(help="Some epsilon", metavar="E", default=0.1)
@@ -886,7 +929,7 @@ def test_metavars_in_help_text(capsys: pytest.CaptureFixture[str]) -> None:
     )
 
 
-@pre_python_10
+@pre_python_3_10
 def test_metavars_in_help_text_not_allowed_for_bool(capsys: pytest.CaptureFixture[str]) -> None:
     class Args(TypedArgs):
         epsilon: bool = arg(metavar="X")
@@ -898,7 +941,7 @@ def test_metavars_in_help_text_not_allowed_for_bool(capsys: pytest.CaptureFixtur
 # Misc
 
 
-@pre_python_10
+@pre_python_3_10
 def test_forwarding_of_argparse_kwargs(capsys: pytest.CaptureFixture[str]) -> None:
     class Args(TypedArgs):
         verbose: bool