+++ tests/test_type_utils.py
@@ -1,6 +1,10 @@
-from typing import List, Optional
+from __future__ import annotations
+
+from typing import List, Optional, Union
 
 from typed_argparse.type_utils import TypeAnnotation, collect_type_annotations
+
+from ._testing_utils import starting_with_python_3_10
 
 # -----------------------------------------------------------------------------
 # TypeAnnotation
@@ -36,6 +40,7 @@ def test_type_annotation__user_class() -> None:
     assert t.validate(12345) == (12345, "value is of type 'int', expected 'MyClass'")
 
 
+@starting_with_python_3_10
 def test_type_annotation__optional() -> None:
     t = TypeAnnotation(Optional[str])
     t_underlying = t.get_underlying_if_optional()
@@ -44,6 +49,33 @@ def test_type_annotation__optional() -> None:
 
     assert t.validate("foo") == ("foo", None)
     assert t.validate(None) == (None, None)
+
+    t = TypeAnnotation(str | None)  # type: ignore[operator, unused-ignore]
+    t_underlying = t.get_underlying_if_optional()
+    assert t_underlying is not None
+    assert t_underlying.raw_type is str
+
+    assert t.validate("foo") == ("foo", None)
+    assert t.validate(None) == (None, None)
+
+
+@starting_with_python_3_10
+def test_type_annotation__union() -> None:
+    t = TypeAnnotation(Union[str, int])
+    t_underlyings = t.get_underlyings_if_union()
+    assert t_underlyings[0] is not None
+    assert t_underlyings[0].raw_type is str
+    assert t_underlyings[1] is not None
+    assert t_underlyings[1].raw_type is int
+
+    t = TypeAnnotation(str | int)  # type: ignore[operator, unused-ignore]
+    t_underlyings = t.get_underlyings_if_union()
+    assert t_underlyings[0] is not None
+    assert t_underlyings[0].raw_type is str
+    assert t_underlyings[1] is not None
+    assert t_underlyings[1].raw_type is int
+
+    # TODO: Check validation semantics
 
 
 def test_type_annotation__list() -> None:
@@ -87,7 +119,7 @@ def test_type_annotation__literals__from_typing_extensions() -> None:
 
 def test_type_annotation__literals__from_typing() -> None:
     # Literal can behave differently whether it comes from typing or typing_extensions
-    from typing import Literal  # type: ignore
+    from typing import Literal
 
     t = TypeAnnotation(Literal[1, 2, 3])